Tryout for Web Database.

+++ scripts/pbedit.js
@@ -34,34 +34,22 @@ function load_image_on_canvas( src )
 		image_layer.image = workspace_image;
 		
 	var workspace = document.getElementById('workspace');
-		workspace.width = workspace_image.width;
-		workspace.height = workspace_image.height;
 
-	//workspace scaling
-	var height_ratio = workspace_wrapper.offsetHeight / workspace_image.height;
-	var width_ratio = workspace_wrapper.offsetWidth / workspace_image.width;
+ 	var long_rate = Math.max( workspace_image.width , workspace_image.height ) / 1200;
+ 	if( long_rate > 1 )
+ 	{
+ 		workspace_image.width = parseInt( workspace_image.width / long_rate ) ;
+ 		workspace_image.height = parseInt( workspace_image.height / long_rate );
+ 	}
+ 		workspace.width = workspace_image.width;
+ 		workspace.height = workspace_image.height;
 
-	//image initial scaling	
-	var workspace_scale = Math.min( height_ratio , width_ratio )
-		workspace.style.height = parseInt( workspace_image.height * workspace_scale - 5) + "px";
-		workspace.style.width = parseInt( workspace_image.width * workspace_scale -5) + "px";
-
-	//scale slider initial set
-	var height_ratio_for_slider = workspace_image.height / workspace_wrapper.offsetHeight;
-	var width_ratio_for_slider = workspace_image.width / workspace_wrapper.offsetWidth;
-
-	//image initial scaling	
-	var slider_scale = Math.min( height_ratio_for_slider , width_ratio_for_slider )
-	
-	var scale_slider = document.getElementById('scale_slider');
-	scale_slider.image_max = slider_scale;
+	fit_scale();
 	
 	ctx = workspace.getContext('2d');
 	ctx.drawImage(workspace_image, 0, 0, workspace.width, workspace.height);
 	
 	var image_data = ctx.getImageData( 0, 0, workspace_image.width , workspace_image.height);
-//	smaller image for debug.
-//	var image_data = ctx.getImageData( 0, 0, 700 , 700);
 
 	init_image_temp();
 	load_image_onto_database( image_data );
@@ -75,7 +63,7 @@ function load_image_onto_database( getImageData )
 	window.console.log('db onsite' , getImageData, image_height, image_width);
 	
 	var startTime = new Date();
-	window.console.log("Import started at ", startTime);
+	window.console.log("Import started at ", pixelData);
 
 	db.transaction(
 		function(tx)
@@ -87,16 +75,11 @@ function load_image_onto_database( getImageData )
 					tx.executeSql( "INSERT INTO PixelData (x , y , R , G, B, A) VALUES (?, ?, ?, ?, ?, ?)" ,
 					[j ,
 					 i ,
-					 pixelData[ j + ( i * image_height )    ] / 255 ,
-					 pixelData[ j + ( i * image_height ) + 1] / 255 ,
-					 pixelData[ j + ( i * image_height ) + 2] / 255 ,
-					 pixelData[ j + ( i * image_height ) + 3] / 255]
-// 					 pixelData[ j + ( i * image_height )    ] ,
-// 					 pixelData[ j + ( i * image_height ) + 1] ,
-// 					 pixelData[ j + ( i * image_height ) + 2] ,
-// 					 pixelData[ j + ( i * image_height ) + 3] ]
+					 pixelData[ ( j + ( i * image_width ) ) * 4    ] / 255 ,
+					 pixelData[ ( j + ( i * image_width ) ) * 4 + 1] / 255 ,
+					 pixelData[ ( j + ( i * image_width ) ) * 4 + 2] / 255 ,
+					 pixelData[ ( j + ( i * image_width ) ) * 4 + 3] / 255]
 					);
-					
 				}
 			}
 			var endTime = new Date();
@@ -120,7 +103,8 @@ function init_web_database(){
 	{
 		window.console.log(error.message);
 	}
-	 init_image_temp();
+	document.first_load = true;
+	init_image_temp();
 }
 
 function init_image_temp(){
@@ -130,20 +114,52 @@ function init_image_temp(){
 		{
 			var exist_table = 0;
 			tx.executeSql(
-				"SELECT count(x) FROM PixelData LIMIT 1" , [] 
+				"SELECT MAX(x), MAX(y) FROM PixelData" , [] ,
 				function(tx, result)
 				{
-					tx.executeSql(
-						"DROP TABLE PixelData",
-						[],
-						function()
-						{
-							window.console.log('Table removed successfully');
-						},
-						function(tx, error)
-						{
-							window.console.log("Error with table removing" , error.message);
-						});
+					if( document.first_load != true )
+					{
+						tx.executeSql(
+							"DROP TABLE PixelData",
+							[],
+							function()
+							{
+								window.console.log('Table removed successfully');
+							},
+							function(tx, error)
+							{
+								window.console.log("Error with table removing" , error.message);
+							});
+						tx.executeSql(
+							"CREATE TABLE PixelData ( x NUMBER, y NUMBER, R REAL, G REAL, B REAL, A REAL)" , [], 
+							function(tx, result)
+							{
+								window.console.log('successed to create table');
+							},
+							function(tx, err)
+							{
+								window.console.log('Error', err);
+							});
+					}
+					else
+					{
+						// restore image from database.
+						var row = result.rows.item(0);
+
+						var image_width = row['MAX(x)'] + 1;
+						var image_height = row['MAX(y)'] + 1;
+						
+						window.console.log(image_width, image_height)
+
+						var workspace = document.getElementById('workspace');
+							workspace.width = image_width;
+							workspace.height = image_height;
+
+						fit_scale();
+						document.first_load = false;
+						//color_evolved_view( [] );
+						redraw();
+					}
 				},
 				function( tx, error ) {
 					tx.executeSql(
@@ -162,19 +178,60 @@ function init_image_temp(){
 	);
 }
 
+function color_evolved_view()
+{
+	db.transaction(
+		function(tx){
+// 			tx.executeSql(
+// 				"SELECT COUNT(x) FROM Filter LIMIT 1",[],
+// 				function(tx){},
+// 				function(tx, err)
+// 				{
+// 					tx.executeSql(
+// 						"DROP VIEW Filter",[]
+// 					);
+// 					window.console.log('removed Filter');
+// 				}
+// 			);
+// 
+			var mySqlSeed = "CREATE VIEW Filter AS SELECT x, y ,R ,G ,B ,A FROM PixelData";
+			var SqlReplace = /R\ \,\G\ \,B\ /;
+			var expression = color_expression();
+			var expression_str = expression[0] + "," + expression[1] + "," + expression[2] ;
+			var exSql = mySqlSeed.replace( SqlReplace , expression_str);
+			tx.executeSql(
+				exSql,
+				[],
+				function(tx, result)
+				{
+					window.console.log( exSql, 'successed to create color Filter');
+				},
+				function(tx, err)
+				{
+					window.console.log('Error', err.message);
+				});
+		}
+	);
+}
+
+function color_expression()
+{
+	var R_biass = "R*" + (( document.getElementById("layer_R").value / 50 ) + 1);
+	var G_biass = "G*" + (( document.getElementById("layer_G").value / 50 ) + 1);
+	var B_biass = "B*" + (( document.getElementById("layer_B").value / 50 ) + 1);
+	
+	window.console.log([ R_biass ,G_biass ,B_biass ]);
+	return([ R_biass ,G_biass ,B_biass ]);
+}
+
 function blur()
 {
 	var canvas =  document.getElementById("workspace");
 	var ctx = canvas.getContext("2d");
 	var image_data = ctx.getImageData( 0, 0, canvas.width , canvas.height);
 
-	var R = 0;
-	var G = 0;
-	var B = 0;
-
-	setTimeout( function(){ preview_process( 100 , image_data) } , 1);
-	
-	setTimeout( function(){ real_process( image_data) } , 500);
+	setTimeout( function(){ preview_process( 10 , image_data) } , 1);
+	setTimeout( function(){ real_process( image_data) } , 50);
 
 	document.imagedata = image_data;
 }
@@ -183,7 +240,7 @@ function redraw()
 {
 //  Sample codes
 //	following code is for workers in future when WebKit supports openDatabase inside workers.
-//
+/*
 // 	window.console.log("redraw start");
 // 	var redraw = new Worker("scripts/worker_redraw.js");
 // 	redraw.postMessage('go');
@@ -192,12 +249,56 @@ function redraw()
 // 	{
 // 		window.console.log(event.data);
 // 	};
-//
+*/
+	color_evolved_view();
+	
+	db.transaction(
+		function(tx)
+		{
+			var canvas =  document.getElementById("workspace");
+			var ctx = canvas.getContext("2d");
 
+			var PixelDataArray = ctx.getImageData(0, 0, canvas.width, canvas.height);
+
+			//var mySqlSeed = "SELECT R ,G ,B ,A FROM Filter";
+			var mySqlSeed = "SELECT R ,G ,B ,A FROM PixelData";
+//			var SqlReplace = /R\ \,\G\ \,B\ /;
+//			var expression = color_expression();
+//			var expression_str = '"' + expression[0] + '","' + expression[1] + '","' + expression[2] + '"' ;
+//			var exSql = mySqlSeed.replace( SqlReplace , expression_str);
+
+//			window.console.log('get filter', exSql);
+
+			tx.executeSql(
+				mySqlSeed,
+				[],
+				function(tx , result)
+				{
+					for( var i = 0 ; i < result.rows.length ; i++ )
+					{
+						var row = result.rows.item(i);
+// 						PixelDataArray.data[ i * 4     ] = parseInt(row['"'+expression[0]+'"'] * 255); // R
+// 						PixelDataArray.data[ i * 4 + 1 ] = parseInt(row['"'+expression[1]+'"'] * 255); // G
+// 						PixelDataArray.data[ i * 4 + 2 ] = parseInt(row['"'+expression[2]+'"'] * 255); // B
+// 						PixelDataArray.data[ i * 4 + 3 ] = parseInt(row['A'] * 255); // A
+						PixelDataArray.data[ i * 4     ] = parseInt(row['R'] * (( document.getElementById("layer_R").value / 50 ) + 1) * 255); // R
+						PixelDataArray.data[ i * 4 + 1 ] = parseInt(row['G'] * (( document.getElementById("layer_G").value / 50 ) + 1) * 255); // G
+						PixelDataArray.data[ i * 4 + 2 ] = parseInt(row['B'] * (( document.getElementById("layer_B").value / 50 ) + 1) * 255); // B
+						PixelDataArray.data[ i * 4 + 3 ] = parseInt(row['A'] * 255); // A
+					}
+					window.console.log('Export', PixelDataArray);
+					ctx.putImageData(PixelDataArray, 0, 0);
+				}
+			);
+		}
+	);
 }
 
 function preview_process( scale , image_data )
 {
+	var canvas =  document.getElementById("workspace");
+	var ctx = canvas.getContext("2d");
+
 	var preArray = new Array();
 	var preR = 0;
 	var preG = 0;
@@ -219,12 +320,18 @@ function preview_process( scale , image_data )
 
 function real_process( image_data )
 {
+	var canvas =  document.getElementById("workspace");
+	var ctx = canvas.getContext("2d");
+
 	for( var i = 0 ; i < (image_data.data.length) ; i ++ )
 	{
 		image_data.data[i]     = index_color ("R" , real_color( "R" , image_data.data[ i ]    ) * 1.1);
 		image_data.data[i + 1] = index_color ("G" , real_color( "G" , image_data.data[ i + 1 ]) * 1.3);
 		image_data.data[i + 2] = index_color ("B" , real_color( "B" , image_data.data[ i + 2 ]) * 1.5);
 	}
+	
+
+	var ctx = canvas.getContext("2d");
 	ctx.putImageData(image_data, 0, 0);
 }
